+++ src/components/folder/LinkBar.tsx
@@ -1,0 +1,69 @@
+import { useState, useRef, FormEvent } from 'react';
+import styled from 'styled-components';
+import BlueBtn from 'components/common/Button/BlueBtn';
+import linkIcon from 'assets/images/link.svg';
+import ModalPortal from 'components/common/Modal/ModalPortal';
+import AddToFolderModal from 'components/common/Modal/AddToFolderModal';
+import useModal from 'hooks/useModal';
+
+function LinkBar() {
+  const { isOpen, handleModalOpen, handleModalClose } = useModal();
+  const [inputValue, setInputValue] = useState('');
+  const input = useRef<HTMLInputElement>(null);
+
+  function handleLinkAdd(event: FormEvent<HTMLFormElement>): void {
+    event.preventDefault();
+    setInputValue(input.current?.value || '');
+    handleModalOpen();
+  }
+
+  return (
+    <>
+      <Container onSubmit={handleLinkAdd}>
+        <Input ref={input} placeholder="링크를 추가해 보세요" />
+        <BlueBtn type="linkAdd">추가하기</BlueBtn>
+      </Container>
+      {isOpen && (
+        <ModalPortal>
+          <AddToFolderModal url={inputValue} onClickClose={() => handleModalClose()} />
+        </ModalPortal>
+      )}
+    </>
+  );
+}
+
+export default LinkBar;
+
+const Container = styled.form`
+  width: 800px;
+  margin: 0 auto;
+  position: relative;
+
+  @media (max-width: 1199px) {
+    width: 100%;
+  }
+`;
+
+const Input = styled.input`
+  width: 100%;
+  height: 69px;
+  padding: 16px 20px 16px 52px;
+  background-image: url(${linkIcon});
+  background-repeat: no-repeat;
+  background-position: 20px 50%;
+  border-radius: 15px;
+  border: 1px solid var(--primary-color);
+  background-color: white;
+  font-size: 1.6rem;
+  outline: none;
+  &::placeholder {
+    color: var(--gray-60);
+  }
+  &:focus {
+    box-shadow: 0px 5px 20px 0px rgba(0, 0, 0, 0.1);
+  }
+  /* Mobile */
+  @media (max-width: 767px) {
+    height: 53px;
+  }
+`;
